czinspect: fix dynamic allocation library.
The number of bytes used in an allocation is now stored within the lzstring
structure, which removes the need for passing length pointers around the place
in scan.c and util.c.

+++ src/czinspect/src/czinspect.c
@@ -203,7 +203,7 @@ static void parse_opt_check(int opt) {
 static void check_ops() {
     uint8_t num = 0;
 
-    for (int i = 0; i < sizeof(uint8_t) * 8; i++)
+    for (int i = 0; i < (int)sizeof(uint8_t) * 8; i++)
         num += (cfg.operation >> i) % 2;
 
     if (num > 1)
